Changed things to get a closer "WKD project  experience" ;)
- xml config
- namings
