--- original-bean/src/main/java/de/dhiller/spring/bugs/MyBean.java
@@ -1,4 +1,0 @@
-package de.dhiller.spring.bugs;
-
-public interface MyBean {
-}
