+++ original-bean/src/main/java/de/dhiller/spring/bugs/original/MyBean.java
@@ -1,0 +1,4 @@
+package de.dhiller.spring.bugs.original;
+
+public interface MyBean {
+}
