+++ original-bean/src/main/java/de/dhiller/spring/bugs/original/OriginalBean.java
@@ -1,0 +1,8 @@
+package de.dhiller.spring.bugs.original;
+
+import javax.inject.Named;
+
+import de.dhiller.spring.bugs.original.MyBean;
+
+public class OriginalBean implements MyBean {
+}
