--- original-bean/src/main/java/de/dhiller/spring/bugs/original_bean/OriginalBean.java
@@ -1,6 +1,0 @@
-package de.dhiller.spring.bugs.original_bean;
-
-import de.dhiller.spring.bugs.MyBean;
-
-public class OriginalBean implements MyBean {
-}
