+++ overridden-bean/src/main/java/de/dhiller/spring/bugs/overridden/Application.java
@@ -1,19 +1,19 @@
-package de.dhiller.spring.bugs;
+package de.dhiller.spring.bugs.overridden;
 
-import de.dhiller.spring.bugs.original_bean.OriginalBean;
-import de.dhiller.spring.bugs.overriding_bean.OverridingBean;
+import de.dhiller.spring.bugs.original.MyBean;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 public class Application {
 
     private static final Logger log = LoggerFactory.getLogger(Application.class);
 
     public static void main(String[] args) {
-        ApplicationContext context = new AnnotationConfigApplicationContext(OverridingConfiguration.class);
-        MyBean obj = (MyBean) context.getBean("myBean");
+        ApplicationContext context = new ClassPathXmlApplicationContext("application-config.xml");
+        MyBean obj = context.getBean(MyBean.class);
         log.info("MyBean is a {}", obj.getClass().getCanonicalName());
         if (! obj.getClass().isAssignableFrom(OverridingBean.class))
             throw new IllegalStateException("Not an overriding bean!");
