+++ overridden-bean/src/main/java/de/dhiller/spring/bugs/overridden/OverridingBean.java
@@ -1,0 +1,6 @@
+package de.dhiller.spring.bugs.overridden;
+
+import de.dhiller.spring.bugs.original.MyBean;
+
+public class OverridingBean implements MyBean {
+}
