--- overridden-bean/src/main/java/de/dhiller/spring/bugs/overriding_bean/OverridingBean.java
@@ -1,6 +1,0 @@
-package de.dhiller.spring.bugs.overriding_bean;
-
-import de.dhiller.spring.bugs.MyBean;
-
-public class OverridingBean implements MyBean {
-}
