+++ pom.xml
@@ -15,6 +15,11 @@
     </properties>
 
     <dependencies>
+    	<dependency>
+    		<groupId>javax.inject</groupId>
+    		<artifactId>javax.inject</artifactId>
+    		<version>1</version>
+    	</dependency>
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-core</artifactId>
